Move firestore fetch to App.js and send as props to AddItem.js
