+++ src/App.js
@@ -1,4 +1,5 @@
 import React, { useEffect, useState } from 'react';
+import { db } from './lib/firebase';
 import { checkLocalStorageForKey } from './lib/localStorage';
 import {
   BrowserRouter as Router,
@@ -14,11 +15,27 @@ import AddItem from './pages/AddItem';
 
 const App = () => {
   const [token, setToken] = useState('');
+  const [listItems, setListItems] = useState([]);
 
   useEffect(() => {
     const retrievedToken = checkLocalStorageForKey('token', '');
     setToken(retrievedToken);
   }, []);
+
+  useEffect(() => {
+    if (token) {
+      db.collection(token)
+        .get()
+        .then((querySnapshot) => {
+          const listItemData = [];
+
+          querySnapshot.forEach((doc) => {
+            listItemData.push(doc.data());
+          });
+          setListItems(listItemData);
+        });
+    }
+  }, [token]);
 
   return (
     <Router>
@@ -28,10 +45,14 @@ const App = () => {
             <Home setToken={setToken} />
           </Route>
           <Route exact path="/list">
-            {token ? <List token={token} /> : <Redirect to="/" />}
+            {token && listItems ? <List token={token} /> : <Redirect to="/" />}
           </Route>
           <Route exact path="/add-item">
-            <AddItem />
+            <AddItem
+              listItems={listItems}
+              setListItems={setListItems}
+              token={token}
+            />
           </Route>
         </Switch>
         {token && <NavBar />}
